+++ .vscode/settings.json
@@ -9,6 +9,7 @@
     "amplify/backend/awscloudformation": true
   },
   "cSpell.words": [
-    "amplifyconfig"
+    "amplifyconfig",
+    "amplifyconfiguration"
   ]
 }
