+++ analysis_options.yaml
@@ -1,29 +1,32 @@
-# This file configures the analyzer, which statically analyzes Dart code to
-# check for errors, warnings, and lints.
-#
-# The issues identified by the analyzer are surfaced in the UI of Dart-enabled
-# IDEs (https://dart.dev/tools#ides-and-editors). The analyzer can also be
-# invoked from the command line by running `flutter analyze`.
-
-# The following line activates a set of recommended lints for Flutter apps,
-# packages, and plugins designed to encourage good coding practices.
 include: package:flutter_lints/flutter.yaml
-
+analyzer:
+  exclude:
+    - lib/amplify_gen/**.dart
+    - lib/amplifyconfiguration.dart
 linter:
-  # The lint rules applied to this project can be customized in the
-  # section below to disable rules from the `package:flutter_lints/flutter.yaml`
-  # included above or to enable additional rules. A list of all available lints
-  # and their documentation is published at
-  # https://dart-lang.github.io/linter/lints/index.html.
-  #
-  # Instead of disabling a lint rule for the entire project in the
-  # section below, it can also be suppressed for a single line of code
-  # or a specific dart file by using the `// ignore: name_of_lint` and
-  # `// ignore_for_file: name_of_lint` syntax on the line or in the file
-  # producing the lint.
   rules:
-    # avoid_print: false  # Uncomment to disable the `avoid_print` rule
-    # prefer_single_quotes: true  # Uncomment to enable the `prefer_single_quotes` rule
+    avoid_print: true
+    prefer_single_quotes: true
+    use_build_context_synchronously: false
+    require_trailing_commas: true
+    implicit_call_tearoffs: true
+    unreachable_from_main: true
+
+    # import
+    prefer_relative_imports: true
+    always_use_package_imports: false
+
+    # 変数の宣言
+    prefer_final_locals: true
+    prefer_const_declarations: true
+    always_declare_return_types: true
+
+    # constructor
+    sort_constructors_first: true
+    sort_unnamed_constructors_first: true
+    prefer_constructors_over_static_methods: true
+    prefer_asserts_in_initializer_lists: true
+    prefer_asserts_with_message: true
 
 # Additional information about this file can be found at
 # https://dart.dev/guides/language/analysis-options
